posts_gatherer/requirements.txt
-i https://pypi.org/simple
boto3==1.16.56
botocore==1.19.56
certifi==2020.12.5
chardet==4.0.0
dnspython==1.16.0
idna==2.10
jmespath==0.10.0
praw==7.1.0
prawcore==1.5.0
pymongo[srv]==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2 ; python_version != '3.4'
websocket-client==0.57.0
